// projects/akita-sc/smart_contracts/gates/plugins/follower-index/types.ts
import { arc4, uint64 } from "@algorandfoundation/algorand-typescript"
import { Operator } from "../../types"
import { Address } from "@algorandfoundation/algorand-typescript/arc4"

export type FollowerIndexRegistryInfo = {
    user: Address
    op: Operator
    value: uint64
}

export class arc4FollowerIndexRegistryInfo extends arc4.Struct<{
    user: Address
    op: arc4.UintN64
    value: arc4.UintN64
}> {}

export type FollowerIndexGateCheckParams = {
    registryID: uint64
    index: uint64
    follower: Address
}

export class arc4FollowerIndexGateCheckParams extends arc4.Struct<{
    registryID: arc4.UintN64
    index: arc4.UintN64
    follower: Address
}> {}